Tidy up code structure && sidechain stroage method
