+++ tee-worker/ts-tests/common/config.js
