+++ tee-worker/ts-tests/common/transactions.ts
@@ -1,0 +1,97 @@
+import { ApiPromise } from "@polkadot/api";
+import { ApiTypes, SubmittableExtrinsic } from "@polkadot/api/types";
+import { TransactionSubmit } from "./type-definitions";
+import { KeyringPair } from '@polkadot/keyring/types';
+import { getListenTimeoutInBlocks } from "./utils";
+import { EventRecord, Event } from "@polkadot/types/interfaces";
+
+//transactions utils
+export async function sendTxUntilInBlock(api: ApiPromise, tx: SubmittableExtrinsic<ApiTypes>, signer: KeyringPair) {
+    return new Promise<{ block: string }>(async (resolve, reject) => {
+        // The purpose of paymentInfo is to check whether the version of polkadot/api is suitable for the current test and to determine whether the transaction is successful.
+        await tx.paymentInfo(signer);
+        const nonce = await api.rpc.system.accountNextIndex(signer.address);
+        await tx.signAndSend(signer, { nonce }, (result) => {
+            if (result.status.isInBlock) {
+                console.log(`Transaction included at blockHash ${result.status.asInBlock}`);
+                resolve({
+                    block: result.status.asInBlock.toString(),
+                });
+            } else if (result.status.isInvalid) {
+                reject(`Transaction is ${result.status}`);
+            }
+        });
+    });
+}
+
+export async function sendTxUntilInBlockList(api: ApiPromise, txs: TransactionSubmit[], signer: KeyringPair) {
+    return Promise.all(
+        txs.map(async ({ tx, nonce }) => {
+            const result = await new Promise((resolve, reject) => {
+                tx.signAndSend(signer, { nonce }, (result) => {
+                    if (result.status.isInBlock) {
+                        //catch error
+                        if (result.dispatchError) {
+                            if (result.dispatchError.isModule) {
+                                const decoded = api.registry.findMetaError(result.dispatchError.asModule);
+                                const { docs, name, section } = decoded;
+
+                                console.log(`${section}.${name}: ${docs.join(' ')}`);
+                                resolve(`${section}.${name}`);
+                            } else {
+                                console.log(result.dispatchError.toString());
+                                resolve(result.dispatchError.toString());
+                            }
+                        } else {
+                            console.log(`Transaction included at blockHash ${result.status.asInBlock}`);
+                            resolve({
+                                block: result.status.asInBlock.toString(),
+                            });
+                        }
+                    } else if (result.status.isInvalid) {
+                        reject(`Transaction is ${result.status}`);
+                    }
+                });
+            });
+            return result;
+        })
+    );
+}
+
+// Subscribe to the chain until we get the first specified event with given `section` and `methods`.
+// We can listen to multiple `methods` as long as they are emitted in the same block.
+// The event consumer should do the decryption optionaly as it's event specific
+export async function listenEvent(api: ApiPromise, section: string, methods: string[]) {
+    return new Promise<Event[]>(async (resolve, reject) => {
+        let startBlock = 0;
+        const unsubscribe = await api.rpc.chain.subscribeNewHeads(async (header) => {
+            const currentBlockNumber = header.number.toNumber();
+            if (startBlock == 0) startBlock = currentBlockNumber;
+            const timeout = await getListenTimeoutInBlocks(api);
+            if (currentBlockNumber > startBlock + timeout) {
+                reject('timeout');
+                return;
+            }
+            console.log(`\n--------- block #${header.number}, hash ${header.hash} ---------\n`);
+            const apiAt = await api.at(header.hash);
+
+            const records: EventRecord[] = (await apiAt.query.system.events()) as any;
+            records.forEach((e, i) => {
+                const s = e.event.section;
+                const m = e.event.method;
+                const d = e.event.data;
+                console.log(`Event[${i}]: ${s}.${m} ${d}`);
+            });
+            const events = records.filter(
+                ({ phase, event }) =>
+                    phase.isApplyExtrinsic && section === event.section && methods.includes(event.method)
+            );
+
+            if (events.length) {
+                resolve(events.map((e) => e.event));
+                unsubscribe();
+                return;
+            }
+        });
+    });
+}
