--- tee-worker/ts-tests/web3/setup.ts
@@ -1,16 +1,0 @@
-import { Keyring } from '@polkadot/api';
-import type { KeyringPair } from '@polkadot/keyring/types';
-const keyring = new Keyring({ type: 'sr25519' });
-const crypto = require('crypto');
-export function getSigner(): KeyringPair[] {
-    let Alice = keyring.addFromUri('//Alice', { name: 'Alice' });
-    let Bob = keyring.addFromUri('//Bob', { name: 'Bob' });
-    let Charlie = keyring.addFromUri('//Charlie', { name: 'Charlie' });
-    let Eve = keyring.addFromUri('//Eve', { name: 'Eve' });
-    const signers = [Alice, Bob, Charlie, Eve];
-    return signers;
-}
-
-export function generateChallengeCode(): String {
-    return crypto.randomBytes(16).toString('hex');
-}
